Move validation schemas
