+++ admin-ui/package.json
@@ -111,6 +111,7 @@
     "@types/styled-components": "^5.1.26",
     "customize-cra": "^1.0.0",
     "eslint-config-airbnb-typescript-prettier": "^5.0.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-testcafe": "^0.2.1",
     "husky": "^8.0.3",
