+++ admin-ui/src/component/my-units/MyUnitRecurringReservation/hooks.ts
@@ -7,8 +7,8 @@ import type {
 } from "common/types/gql-types";
 import { ReservationUnitsReservationUnitReservationStartIntervalChoices } from "common/types/gql-types";
 import type { UseFormReturn } from "react-hook-form";
+import type { RecurringReservationForm } from "app/schemas";
 import { generateReservations } from "./generateReservations";
-import type { RecurringReservationForm } from "./RecurringReservationSchema";
 import { useNotification } from "../../../context/NotificationContext";
 import { RECURRING_RESERVATION_UNIT_QUERY } from "../queries";
