+++ admin-ui/src/component/my-units/create-reservation/utils.ts
@@ -1,7 +1,8 @@
 import { get, pick, zipObject } from "lodash";
-import { ReservationUnitsReservationUnitReservationStartIntervalChoices } from "common/types/gql-types";
-import type { ReservationFormType } from "./validator";
-import type { RecurringReservationForm } from "../MyUnitRecurringReservation/RecurringReservationSchema";
+import {
+  type ReservationFormType,
+  type RecurringReservationForm,
+} from "app/schemas";
 
 export function flattenMetadata(
   values: ReservationFormType | RecurringReservationForm,
@@ -18,20 +19,3 @@ export function flattenMetadata(
     Object.values(metadataSetValues).map((v) => get(v, "value") || v)
   );
 }
-
-export const intervalToNumber = (
-  i: ReservationUnitsReservationUnitReservationStartIntervalChoices
-) => {
-  switch (i) {
-    case ReservationUnitsReservationUnitReservationStartIntervalChoices.Interval_15Mins:
-      return 15;
-    case ReservationUnitsReservationUnitReservationStartIntervalChoices.Interval_30Mins:
-      return 30;
-    case ReservationUnitsReservationUnitReservationStartIntervalChoices.Interval_60Mins:
-      return 60;
-    case ReservationUnitsReservationUnitReservationStartIntervalChoices.Interval_90Mins:
-      return 90;
-    default:
-      return 0;
-  }
-};
