+++ admin-ui/src/schemas/recurringReservation.test.ts
@@ -3,7 +3,7 @@ import { expect, test } from "@jest/globals";
 import { addDays } from "date-fns";
 
 import { ReservationUnitsReservationUnitReservationStartIntervalChoices } from "common/types/gql-types";
-import { timeSelectionSchema } from "./RecurringReservationSchema";
+import { timeSelectionSchema } from "./recurringReservation";
 
 const tomorrow = addDays(new Date(), 1);
 const reservation = {
