+++ admin-ui/src/schemas/reservation.test.ts
@@ -2,8 +2,8 @@ import { ReservationUnitsReservationUnitReservationStartIntervalChoices } from "
 import { addDays, addHours, format, setMinutes, subDays } from "date-fns";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { expect, test, describe } from "@jest/globals";
-import { formatDate } from "../../../common/util";
-import { ReservationFormSchema } from "./validator";
+import { ReservationFormSchema } from "./reservation";
+import { formatDate } from "../common/util";
 
 const TIME_FORMAT = "HH:mm";
